requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
tensorflow>=2.13.0
torch>=2.0.0
xgboost>=2.0.0
catboost>=1.2.0
lightgbm
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Technical Analysis

# Trading & Exchange
python-binance>=1.0.17
ccxt>=4.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Async & Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Web Dashboard
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Logging & Monitoring
loguru>=0.7.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

fastapi
uvicorn
requests